gh-108777: Split _PyTime tests from _testinternalcapi.c
