+++ Modules/_testinternalcapi/parts.h
@@ -1,0 +1,15 @@
+#ifndef Py_TESTINTERNALCAPI_PARTS_H
+#define Py_TESTINTERNALCAPI_PARTS_H
+
+// Always enable assertions
+#undef NDEBUG
+
+#ifndef Py_BUILD_CORE_BUILTIN
+#  define Py_BUILD_CORE_MODULE 1
+#endif
+
+#include "Python.h"
+
+int _PyTestInternalCapi_Init_PyTime(PyObject *module);
+
+#endif // Py_TESTINTERNALCAPI_PARTS_H
