+++ PCbuild/_testinternalcapi.vcxproj.filters
@@ -12,6 +12,9 @@
     <ClCompile Include="..\Modules\_testinternalcapi.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\Modules\_testinternalcapi\pytime.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\PC\python_nt.rc">
